+++ App.tsx
@@ -1,10 +1,7 @@
-import { createStackNavigator } from "@react-navigation/stack";
 import AppNavigator from "./navigation/AppNavigator";
 
-const Stack = createStackNavigator();
-
 const App = () => {
-  return <AppNavigator></AppNavigator>;
+  return <AppNavigator/>;
 };
 
 export default App;
