+++ navigation/AppTopBar.tsx
@@ -1,9 +1,9 @@
-import * as React from "react";
 import { IconButton, Icon, HStack, Text } from "native-base";
 import { MaterialIcons } from "@expo/vector-icons";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 
 import { useNavigation, useRoute } from "@react-navigation/native";
+import { SIZES } from "../constants/theme";
 
 type RootStackParamList = {
   Home: undefined;
@@ -20,16 +20,34 @@ const AppBar: React.FC<AppBarProps> = () => {
 
   const canGoBack = navigation.canGoBack();
 
-  let headerTitle =
-    route.name === "Impressum" ? (
-      <Text fontSize="20" fontWeight="bold">
-        Impressum
-      </Text>
-    ) : (
-      <Text fontSize="20" fontWeight="bold">
-        BRE<Text style={{ color: "aqua" }}>T</Text>A
-      </Text>
-    );
+  let headerTitle;
+
+if (route.name === "Impressum") {
+  headerTitle = (
+    <Text fontSize={SIZES.large} fontWeight="bold">
+      Impressum
+    </Text>
+  );
+} else if (route.name === "Authentification") {
+  headerTitle = (
+    <Text fontSize={SIZES.large} fontWeight="bold">
+      Authentification
+    </Text>
+  );
+} else if (route.name === "ConfirmCode") {
+  headerTitle = (
+    <Text fontSize={SIZES.large} fontWeight="bold">
+      Login
+    </Text>
+  );
+} else {
+  headerTitle = (
+    <Text fontSize="20" fontWeight="bold">
+      BRE<Text style={{ color: "aqua" }}>T</Text>A
+    </Text>
+  );
+}
+
 
   return (
     <>
@@ -51,9 +69,6 @@ const AppBar: React.FC<AppBarProps> = () => {
         </HStack>
 
         <HStack flex={2} justifyContent="center">
-          {/* <Text fontSize="20" fontWeight="bold">
-            BRE<Text style={{ color: "aqua" }}>T</Text>A
-            </Text> */}
           {headerTitle}
         </HStack>
 
@@ -62,7 +77,7 @@ const AppBar: React.FC<AppBarProps> = () => {
             icon={
               <Icon as={MaterialIcons} name="info" size="sm" color="black" />
             }
-            onPress={() => navigation.navigate("Impressum")}
+            onPress={() => navigation.navigate("Impressum" as never)}
           />
         </HStack>
       </HStack>
