+++ screens/Post.tsx
@@ -1,5 +1,6 @@
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
+import { COLORS } from "../constants/theme";
 
 export default function Post() {
   return (
@@ -13,7 +14,7 @@ export default function Post() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "#fff",
+    backgroundColor: COLORS.white,
     alignItems: "center",
     justifyContent: "center",
   },
